pyproject.toml
[project]
name = "django-task-application"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "djangorestframework>=3.17.1",
    "psycopg2-binary>=2.9.11",
]
